Centralize unit adding and stripping logic to a singular set of functions
