+++ gwcs/tests/test_coordinate_systems.py
@@ -646,6 +646,12 @@ def test_CoordinateFrameProtocol():
         def remove_units(self, arrays):
             return arrays
 
+        def to_high_level_coordinates(self, *inputs):
+            return inputs
+
+        def from_high_level_coordinates(self, *inputs):
+            return inputs
+
     frame = MyFrame()
     assert isinstance(frame, cf.CoordinateFrameProtocol)
